chore: migrate PreCheckMeta/PreCheckMandatoryFields hook to validator
