+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/hooks/EventPreCheckMetaValidator.java
@@ -27,37 +27,42 @@
  */
 package org.hisp.dhis.tracker.validation.hooks;
 
-import static org.hisp.dhis.tracker.validation.hooks.AssertTrackerValidationReport.assertHasError;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1010;
+import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1011;
+import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1013;
 
-import org.hisp.dhis.tracker.TrackerType;
-import org.hisp.dhis.tracker.report.TrackerErrorCode;
+import org.hisp.dhis.organisationunit.OrganisationUnit;
+import org.hisp.dhis.program.Program;
+import org.hisp.dhis.program.ProgramStage;
+import org.hisp.dhis.tracker.bundle.TrackerBundle;
+import org.hisp.dhis.tracker.domain.Event;
 import org.hisp.dhis.tracker.validation.ValidationErrorReporter;
+import org.hisp.dhis.tracker.validation.Validator;
+import org.springframework.stereotype.Component;
 
-public class AssertValidationErrorReporter
+/**
+ * @author Morten Svanæs <msvanaes@dhis2.org>
+ */
+@Component
+public class EventPreCheckMetaValidator
+    implements Validator<Event>
 {
-    private AssertValidationErrorReporter()
+    @Override
+    public void validate( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
     {
-        throw new IllegalStateException( "utility class" );
+        OrganisationUnit organisationUnit = bundle.getPreheat().getOrganisationUnit( event.getOrgUnit() );
+        reporter.addErrorIfNull( organisationUnit, event, E1011, event.getOrgUnit() );
+
+        Program program = bundle.getPreheat().getProgram( event.getProgram() );
+        reporter.addErrorIfNull( program, event, E1010, event.getProgram() );
+
+        ProgramStage programStage = bundle.getPreheat().getProgramStage( event.getProgramStage() );
+        reporter.addErrorIfNull( programStage, event, E1013, event.getProgramStage() );
     }
 
-    public static void assertMissingProperty( ValidationErrorReporter reporter, TrackerType type, String entity,
-        String uid,
-        String property,
-        TrackerErrorCode errorCode )
+    @Override
+    public boolean skipOnError()
     {
-        assertHasError( reporter.getErrors(), errorCode, type, uid,
-            "Missing required " + entity + " property: `" + property + "`." );
-    }
-
-    public static void hasTrackerError( ValidationErrorReporter reporter, TrackerErrorCode code, TrackerType type,
-        String uid )
-    {
-        assertTrue( reporter.hasErrors(), "error not found since reporter has no errors" );
-        assertTrue( reporter.hasErrorReport( err -> code == err.getErrorCode() &&
-            type == err.getTrackerType() &&
-            uid.equals( err.getUid() ) ),
-            String.format( "error with code %s, type %s, uid %s not found in reporter with %d error(s)", code, type,
-                uid, reporter.getErrors().size() ) );
+        return true;
     }
 }
