+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/hooks/TrackedEntityPreCheckMetaValidator.java
@@ -27,37 +27,43 @@
  */
 package org.hisp.dhis.tracker.validation.hooks;
 
-import static org.hisp.dhis.tracker.validation.hooks.AssertTrackerValidationReport.assertHasError;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1005;
 
-import org.hisp.dhis.tracker.TrackerType;
+import org.hisp.dhis.organisationunit.OrganisationUnit;
+import org.hisp.dhis.trackedentity.TrackedEntityType;
+import org.hisp.dhis.tracker.bundle.TrackerBundle;
+import org.hisp.dhis.tracker.domain.TrackedEntity;
 import org.hisp.dhis.tracker.report.TrackerErrorCode;
 import org.hisp.dhis.tracker.validation.ValidationErrorReporter;
+import org.hisp.dhis.tracker.validation.Validator;
+import org.springframework.stereotype.Component;
 
-public class AssertValidationErrorReporter
+/**
+ * @author Morten Svanæs <msvanaes@dhis2.org>
+ */
+@Component
+public class TrackedEntityPreCheckMetaValidator
+    implements Validator<TrackedEntity>
 {
-    private AssertValidationErrorReporter()
+    @Override
+    public void validate( ValidationErrorReporter reporter, TrackerBundle bundle, TrackedEntity tei )
     {
-        throw new IllegalStateException( "utility class" );
+        OrganisationUnit organisationUnit = bundle.getPreheat().getOrganisationUnit( tei.getOrgUnit() );
+        if ( organisationUnit == null )
+        {
+            reporter.addError( tei, TrackerErrorCode.E1049, tei.getOrgUnit() );
+        }
+
+        TrackedEntityType entityType = bundle.getPreheat().getTrackedEntityType( tei.getTrackedEntityType() );
+        if ( entityType == null )
+        {
+            reporter.addError( tei, E1005, tei.getTrackedEntityType() );
+        }
     }
 
-    public static void assertMissingProperty( ValidationErrorReporter reporter, TrackerType type, String entity,
-        String uid,
-        String property,
-        TrackerErrorCode errorCode )
+    @Override
+    public boolean skipOnError()
     {
-        assertHasError( reporter.getErrors(), errorCode, type, uid,
-            "Missing required " + entity + " property: `" + property + "`." );
-    }
-
-    public static void hasTrackerError( ValidationErrorReporter reporter, TrackerErrorCode code, TrackerType type,
-        String uid )
-    {
-        assertTrue( reporter.hasErrors(), "error not found since reporter has no errors" );
-        assertTrue( reporter.hasErrorReport( err -> code == err.getErrorCode() &&
-            type == err.getTrackerType() &&
-            uid.equals( err.getUid() ) ),
-            String.format( "error with code %s, type %s, uid %s not found in reporter with %d error(s)", code, type,
-                uid, reporter.getErrors().size() ) );
+        return true;
     }
 }
